Remove linter errors
